finished
The website is in a "finished" state now. When the user is logged in there is now a "load old trips" button that, if pressed at all, has to be pressed before "create new trip". It loads all of the trips from the database for that user. Do not save a trip twice, it will save with the same key. You can also press "load weather" on any trip to load the weather without saving it. You can press "load weather" multiple times. Also, the username is implemented correctly. So different users will see different trips depending on what is saved to the db.sqlite file.